backend/requirements.txt
# Context-Aware Digital Twin Email Assistant
# Python Dependencies

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Environment Management
python-dotenv==1.0.0

# Data Processing
python-dateutil==2.8.2

# Future: Gmail API Integration
# google-auth==2.23.0
# google-auth-oauthlib==1.1.0
# google-auth-httplib2==0.1.1
# google-api-python-client==2.100.0

# Future: Machine Learning
# scikit-learn==1.3.1
# numpy==1.26.0
# pandas==2.1.1

# Future: OpenAI API
# openai==1.3.0

# Future: Database
# pymongo==4.5.0
# motor==3.3.1  # Async MongoDB driver

# Development Tools
# pytest==7.4.2
# black==23.9.1
# flake8==6.1.0
